adding command handled
