+++ src/service/Worker.php
@@ -65,6 +65,13 @@ class Worker extends Singleton
         // настройка сети
         if (isset($config['net'])) {
             $this->sc = new TcpClient($config['net']);
+
+            $this->sc->onRead($this->onRead());
+
+            $this->sc->onDisconnect(function () {
+                Log::log('i disconnected!');
+                $this->work = false;
+            });
         } else {
             trigger_error('Net configuration not found', E_USER_NOTICE);
             return false;
@@ -83,6 +90,9 @@ class Worker extends Singleton
             return false;
         }
         require_once $this->worker_app;
+        if (!class_exists($this->worker_app_class)) {
+            trigger_error('Worker application must be configured with "worker_app_class"', E_USER_ERROR);
+        }
         $this->app = new $this->worker_app_class();
         if (!$this->app instanceof Application) {
             trigger_error('Worker application must be instance of maestroprog\saw\Application', E_USER_ERROR);
@@ -97,44 +107,6 @@ class Worker extends Singleton
 
     public function connect()
     {
-        $this->sc->onRead(function ($data) {
-            Log::log('I RECEIVED ' . $data . ' :)');
-
-            switch ($data) {
-                case 'HELLO':
-                    $this->sc->send('HELLO');
-                    break;
-                case 'ACCEPT':
-                    $this->dispatcher
-                        ->create(WorkerAdd::NAME, $this->sc)
-                        ->setError(function () {
-                            $this->stop();
-                        })
-                        ->setSuccess(function () {
-                            //todo
-                        })
-                        ->run();
-                    break;
-                case 'INVALID':
-                    // todo
-                    break;
-                case 'BYE':
-                    $this->work = false;
-                    break;
-                default:
-                    if (is_array($data) && $this->dispatcher->valid($data)) {
-                        $this->handle($data);
-                    } else {
-                        $this->sc->send('INVALID');
-                    }
-            }
-        });
-
-        $this->sc->onDisconnect(function () {
-            Log::log('i disconnected!');
-            $this->work = false;
-        });
-
         return $this->sc->connect();
     }
 
@@ -206,6 +178,42 @@ class Worker extends Singleton
         $this->app->run($this->task);
     }
 
+    protected function onRead()
+    {
+        return function ($data) {
+            Log::log('I RECEIVED ' . $data . ' :)');
+
+            switch ($data) {
+                case 'HELLO':
+                    $this->sc->send('HELLO');
+                    break;
+                case 'ACCEPT':
+                    $this->dispatcher
+                        ->create(WorkerAdd::NAME, $this->sc)
+                        ->setError(function () {
+                            $this->stop();
+                        })
+                        ->setSuccess(function () {
+                            //todo
+                        })
+                        ->run();
+                    break;
+                case 'INVALID':
+                    // todo
+                    break;
+                case 'BYE':
+                    $this->work = false;
+                    break;
+                default:
+                    if (is_array($data) && $this->dispatcher->valid($data)) {
+                        $this->handle($data);
+                    } else {
+                        $this->sc->send('INVALID');
+                    }
+            }
+        };
+    }
+
     protected function handle(array $data)
     {
         try {
@@ -234,6 +242,29 @@ class Worker extends Singleton
             case 'wdel': // controller has been delete this worker from self stack
                 $this->stop();
                 break;
+        }
+    }
+    /**
+     * @param array $config
+     * @return Worker
+     * @throws \Exception
+     */
+    public static function create(array $config) : Worker
+    {
+        $init = Worker::getInstance();
+        if ($init->init($config)) {
+            Log::log('configured. input...');
+            if (!($init->connect())) {
+                Log::log('Worker start failed');
+                throw new \Exception('Worker starting fail');
+            }
+            register_shutdown_function(function () use ($init) {
+                $init->stop();
+                Log::log('closed');
+            });
+            return $init->setTask(Factory::getInstance()->createTask($init));
+        } else {
+            throw new \Exception('Cannot initialize Worker');
         }
     }
 }
